requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
jinja2==2.9.6
markupsafe==1.0           # via jinja2
pyyaml==3.12
